+++ README.md
@@ -54,8 +54,7 @@
 
 가장 마음에 안드는것은 components 폴더다. components 폴더는 어느곳에서 사용할 수 있는 곳인데. 작업을하면 할 수록 공통적인게 아닌 페이지를 위한 폴더가 되어가고 있었다.
 
-아직 페이지도 두개밖에 없어서 상관은 없었지만 나중에는 못할 것 같았다. 그래서 내가 적용한 방법은 features라는 폴더를 하나만들어
-페이지 기능단위의 컴포넌트들을 따로 관리할 수 있게 만들고,
+아직 페이지도 두개밖에 없어서 상관은 없었지만 나중에는 못할 것 같았다. 그래서 내가 적용한 방법은 features라는 폴더를 하나만들어 페이지 기능단위의 컴포넌트들을 따로 관리할 수 있게 만들고,
 
 context, hooks와 같은 특정 페이지에서만 사용할 수 있는 기능들을 같이 만들 수 있게 적용했다. 이렇게 하니까 components 폴더에는 어디서든 사용할 수 있는 즉 재사용성이 높은 컴포넌트들만 남아있었다.
