+++ src/components/article/ArticleList.jsx
@@ -1,20 +1,20 @@
 import React from "react"
-import CardItem from "./CardItem"
-import * as S from "./CardList.style"
+import CardItem from "./ArticleItem"
+import * as S from "./ArticleList.style"
 import AlertBanner from "components/UI/AlertBanner"
 
-function CardList({ data, alertMessage }) {
+function ArticleList({ data, alertMessage }) {
   if (data?.data.length === 0) {
     return <AlertBanner type="info">{alertMessage}</AlertBanner>
   }
 
   return (
-    <S.CardList>
+    <S.ArticleList>
       {data?.data.map((link) => (
         <CardItem key={link.id} data={link} />
       ))}
-    </S.CardList>
+    </S.ArticleList>
   )
 }
 
-export default CardList
+export default ArticleList
