+++ src/components/article/ArticleList.style.js
@@ -1,6 +1,6 @@
 import styled from "styled-components"
 
-export const CardList = styled.ul`
+export const ArticleList = styled.ul`
   display: grid;
   grid-template-columns: repeat(3, 340px);
   gap: 24px 20px;
